data_preparation/requirements.txt
pillow
numpy
torch
torchvision
transformers
umap-learn
